Better HTTP error handling in organization data source
Currently, many forms of misconfiguration result in nil pointer derefs.
Ideally, all HTTP client interaction should be centralized and look
something like this, but that is a very invasive change that should
wait until we have good test coverage in place. However, the organization
data source is very likely to be the first HTTP request made from a
configuration, we fix that to provide proper error messages to the
user.